refactor(fe/NoteReactionList): move ParticipantExtendedInfo logic to NoteReactionList
